SimpleNavigator implements Navigator  setSpeed() takes  an int   parameter   as specified in the Navigator interface
git-svn-id: http://svn.code.sf.net/p/lejos/code/trunk/classes@2224 06c8f799-c779-4783-8af7-179afedb67c8

+++ lejos/navigation/Pilot.java
@@ -39,7 +39,7 @@ public interface Pilot
    * sets drive motor speed deg/sec
    * @param speed
    */
-  public void setSpeed(float speed);
+  public void setSpeed(int speed);
 
  /**
    * Moves the NXT robot a specific distance. A positive value moves it forward and
